new structre moved

+++ Entities/Concrete/Infos/CoverCap.cs
@@ -6,7 +6,7 @@ namespace Entities.Concrete.Infos
     public class CoverCap : IEntity
     {   // kitap kapağı tipi, karton, deri,  vs vs
         [Key]
-        public int Id { get; set; }
+        public byte Id { get; set; }
 
         public string BookSkinType { get; set; }
